Fixed track Duration and PlaylistItem representation in JSON

+++ PlaylistService/PlaylistService/DTOs/TrackDTOs.cs
@@ -1,4 +1,4 @@
 ﻿namespace PlaylistService.DTOs.TrackDTOs
 {
-    public record GetTrackDTO(Guid Id, string Title, string Description, long Duration, string ArtworkUrl, string UrlId, DateTimeOffset UploadDate);
+    public record GetTrackDTO(Guid Id, string Title, string Description, double Duration, string ArtworkUrl, string UrlId, DateTimeOffset UploadDate);
 }
